working on payorders

+++ app/src/main/java/kk/domoRolls/ru/domain/model/PromoCode.kt
@@ -3,7 +3,7 @@ package kk.domoRolls.ru.domain.model
 data class PromoCode(
     val value:String = "",
     val discount: Double = 0.0,
-    val minPrice: Double = 0.0,
+    val minPrice: Int = 0,
     val expirationDate: String = "",
     val isOneTime: Boolean = false
 )
